third_party/WebKit/LayoutTests/xr/xrSession_requestFrameOfReference.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../external/wpt/resources/chromium/webxr-test.js"></script>
<script src="../xr/resources/xr-test-utils.js"></script>
<script src="../xr/resources/test-constants.js"></script>
<canvas id="webgl-canvas"></canvas>

<script>
let testName = "requestFrameOfReference returns the expected objects";

let fakeDeviceInitParams = { supportsImmersive:true };

let requestSessionOptions = [
  { outputContext: getOutputContext() },
  { immersive: true },
];

let testFunction = function(session, t) {
  return promise_rejects(t, new TypeError(), session.requestFrameOfReference("foo"))
    .then(() => Promise.all([
      session.requestFrameOfReference("head-model").then( (frameOfRef) => {
        assert_true(frameOfRef instanceof XRCoordinateSystem,
          "head-model frameOfRef is not correct type.");
        assert_true(frameOfRef instanceof XRFrameOfReference,
          "head-model frameOfRef is not correct type.");
      }),
      session.requestFrameOfReference("eye-level").then( (frameOfRef) => {
        assert_true(frameOfRef instanceof XRCoordinateSystem,
          "eye-level frameOfRef is not correct type.");
        assert_true(frameOfRef instanceof XRFrameOfReference,
          "eye-level frameOfRef is not correct type.");
      }),
      session.requestFrameOfReference("stage").then( (frameOfRef) => {
        assert_true(frameOfRef instanceof XRCoordinateSystem,
          "stage frameOfRef is not correct type.");
        assert_true(frameOfRef instanceof XRFrameOfReference,
          "stage frameOfRef is not correct type.");
      })
  ]));
};

xr_session_promise_test(
  testFunction, fakeDeviceInitParams, requestSessionOptions, testName);

</script>
